requirements.txt
aiogram==3.2.0
aiohttp==3.9.1
PyGithub==2.1.1
